rework use of session storage
