+++ bids-validator/utils/files/readDir.js
@@ -4,29 +4,9 @@ import path from 'path'
 import fs from 'fs'
 import * as child_proccess from 'child_process'
 import isNode from '../isNode'
+import getSessionStorage from '../getSessionStorage'
 
-function createSessionStorageMock() {
-  const sessionStorage = {}
-
-  return {
-    getItem: (key) => sessionStorage[key],
-    setItem: (key, value) => {
-      sessionStorage[key] = value
-    },
-    removeItem: (key) => {
-      delete sessionStorage[key]
-    },
-    clear: () => {
-      Object.keys(sessionStorage).forEach((key) =>
-        sessionStorage.removeItem(key),
-      )
-    },
-  }
-}
-
-const sessionStorage = isNode
-  ? createSessionStorageMock()
-  : window.sessionStorage
+const sessionStorage = isNode ? getSessionStorage() : window.sessionStorage
 
 /**
  * Read Directory
